Add a key event listener to open Preview pane on GitHub comments
Use Apple-Option-P to open the preview pane and toggle it closed.

GitHub uses Apple-Shift-P by default to toggle the Preview pane. I need
this new shortcut because Firefox uses Apple-Shift-P to open new private
browsing windows, shadowing it from GitHub.

This is the initial working version of this script. It doesn't currently
support comment fields that are added to the page after page load.