Enable listener for all comment fields including future ones
The keyboard shortcut now works on all comment fields, including those
that are created after page load.

Thanks to this Stack Overflow answer from 'adeneo' and Suraj Jain:
https://stackoverflow.com/questions/20330945/how-to-addeventlistener-to-future-dom-elements/20331016#20331016

It takes advantage of the fact that JavaScript events bubble. We attach
an event to `<body>`, which we can be sure will always be there (even
across GitHub's Turbolinks), and in the keydown event listener check to
see whether we were called on the right element.